postgresqlserver: Keep track of the polling URL when creating
Previously this was being dropped on the floor unless the create call
returned an AsyncOpIncompleteError. This works ok if the spec is
correct and the server eventually shows up, but for some errors (like
name clashes) the create server call returns no error but querying the
poll URL returns an error about the name clash. Without capturing the
poll URL on the initial creation the reconciliation gets stuck because
it thinks that the 404 it gets when checking for the server indicates
that it's just not ready yet - provisioning never goes back to false
and it doesn't retry the creation.

This is the same fix as that made against mysqlserver in
2f68b47d257d600faa50a1af10d14739f4fc8bdd.